removed sentiment from api and updated so q_type not needed
