+++ api/api.py
@@ -22,10 +22,6 @@ tags_metadata = [
         "name": "multilabel",
         "description": "Generate multilabel predictions for given text.",
     },
-    {
-        "name": "sentiment",
-        "description": "Generate predicted sentiment for given text.",
-    },
 ]
 
 
@@ -57,48 +53,42 @@ def index():
     tags=["multilabel"],
 )
 async def predict_multilabel(items: List[schemas.ItemIn]):
-    """Accepts comment ids, comment text and question type as JSON in a POST request. Makes predictions using trained SVC model.
+    """Accepts comment ids and comment text as JSON in a POST request. Makes predictions using trained SVC model.
 
     Args:
         items (List[ItemIn]): JSON list of dictionaries with the following compulsory keys:
         - `comment_id` (str)
         - `comment_text` (str)
-        - `question_type` (str)
-        The 'question_type' must be one of three values: 'nonspecific', 'what_good', and 'could_improve'.
-        For example, `[{'comment_id': '1', 'comment_text': 'Thank you', 'question_type': 'what_good'},
-        {'comment_id': '2', 'comment_text': 'Food was cold', 'question_type': 'could_improve'}]`
 
     Returns:
-        (dict): Keys are: `comment_id`, `comment_text`, and predicted `labels`.
+        (dict): Keys are: `comment_id` and predicted `labels`.
     """
 
     # Process received data
     df = pd.DataFrame([i.dict() for i in items], dtype=str)
-    df_newindex = df.set_index("comment_id")
-    if df_newindex.index.duplicated().sum() != 0:
-        raise ValueError("comment_id must all be unique values")
-    df_newindex.index.rename("Comment ID", inplace=True)
-    text_to_predict = df_newindex[["comment_text", "question_type"]]
-    text_to_predict = text_to_predict.rename(
-        columns={"comment_text": "FFT answer", "question_type": "FFT_q_standardised"}
+    df_for_preds = df.copy().rename(
+        columns={"comment_id": "Comment ID", "comment_text": "FFT answer"}
     )
+    df_for_preds = df_for_preds.set_index("Comment ID")
+    if df_for_preds.index.duplicated().sum() != 0:
+        raise ValueError("comment_id must all be unique values")
+    text_to_predict = df_for_preds["FFT answer"]
     # Make predictions
-    model_path = "svc_minorcats_v5.sav"
+    model_path = "final_svc.sav"
     if not os.path.isfile(model_path):
         model_path = os.path.join("api", model_path)
     with open(model_path, "rb") as model:
         loaded_model = pickle.load(model)
     preds_df = predict_multilabel_sklearn(
-        text_to_predict, loaded_model, labels=minor_cats, additional_features=True
+        text_to_predict, loaded_model, labels=minor_cats, additional_features=False
     )
     # Join predicted labels with received data
     preds_df["comment_id"] = preds_df.index.astype(str)
     merged = pd.merge(df, preds_df, how="left", on="comment_id")
     merged["labels"] = merged["labels"].fillna("").apply(list)
     for i in merged["labels"].index:
-        if len(merged["labels"].loc[i]) < 1:
-            merged["labels"].loc[i].append("Labelling not possible")
-    return_dict = merged[["comment_id", "comment_text", "labels"]].to_dict(
-        orient="records"
-    )
+        label_list = merged.loc[i, "labels"]
+        if len(label_list) < 1:
+            merged.loc[i, "labels"].append("Labelling not possible")
+    return_dict = merged[["comment_id", "labels"]].to_dict(orient="records")
     return return_dict
